move equals and hash functors to use them in join
