+++ ydb/library/yql/dq/comp_nodes/dq_hash_combine.cpp
@@ -25,49 +25,6 @@ using NUdf::TUnboxedValuePod;
 
 namespace {
 
-struct TWideUnboxedEqual
-{
-    TWideUnboxedEqual(const TKeyTypes& types)
-        : Types(types)
-    {
-    }
-
-    bool operator()(const NUdf::TUnboxedValuePod* left, const NUdf::TUnboxedValuePod* right) const {
-        for (ui32 i = 0U; i < Types.size(); ++i)
-            if (CompareValues(Types[i].first, true, Types[i].second, left[i], right[i]))
-                return false;
-        return true;
-    }
-
-    const TKeyTypes& Types;
-};
-
-struct TWideUnboxedHasher
-{
-    TWideUnboxedHasher(const TKeyTypes& types)
-        : Types(types)
-    {
-    }
-
-    NUdf::THashType operator()(const NUdf::TUnboxedValuePod* values) const {
-        if (Types.size() == 1U)
-            if (const auto v = *values)
-                return NUdf::GetValueHash(Types.front().first, v);
-            else
-                return HashOfNull;
-
-        NUdf::THashType hash = 0ULL;
-        for (const auto& type : Types) {
-            if (const auto v = *values++)
-                hash = CombineHashes(hash, NUdf::GetValueHash(type.first, v));
-            else
-                hash = CombineHashes(hash, HashOfNull);
-        }
-        return hash;
-    }
-
-    const TKeyTypes& Types;
-};
 
 using TEqualsPtr = bool(*)(const NUdf::TUnboxedValuePod*, const NUdf::TUnboxedValuePod*);
 using THashPtr = NUdf::THashType(*)(const NUdf::TUnboxedValuePod*);
